Remove parser after applying a boolean operator
